requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
crontab==0.23.0
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-flower==1.0.0
django-jinja==2.9.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
export==0.1.2
flower==1.0.0
gevent==21.1.2
greenlet==1.1.0
humanize==0.5.1
idna==3.2
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
plumbum==1.7.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
rpyc==5.0.1
six==1.16.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0